Whoami and Whois command refactorized
